Intelligent insert into text field when mode=gui...
... e,g, in databasejoin element advanced tab for inserting description
fields.

1. Position caret at end when initialising
2. Add whitespace as needed around value to avoid concatenation with
existing strings.

This is not a complete fix as e.g. if user positions caret in the middle
of an element name, then inserting a new element name will split the
existing one. This should probably be an HTML list field rather than a
text field, but this is still better than existing.