Various fixes
1.  Remove name attributes from modal popup fields - these are
consolidated into a single field when you close the popup, so the
individual popup fields should not be sent to the server. I have
confirmed that any values previously stored in the DB in error are not
used anywhere in Fabrik.

2. Apply chosen to first row of modal popup as well as to cloned rows.

3. Clone ids so we can change them - no idea why this ever worked
prweviously - but this may be why it didn't save repeated groups
properly.